# queryrepository/src/main/java/fau/dataMart/Controller/ExecutionplanApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.0.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package fau.dataMart.Controller;

import fau.dataMart.dto.ExecutionplanStep;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-02-05T22:08:28.313274270+01:00[Europe/Berlin]")
@Validated
@Tag(name = "query", description = "the query API")
public interface ExecutionplanApi {

    default ExecutionplanApiDelegate getDelegate() {
        return new ExecutionplanApiDelegate() {};
    }

    /**
     * GET /executionplan/{hash}/steps : Get Executionplan and Steps
     *
     * @param hash Get steps of Executionplan (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "executionplanHashStepsGet",
        summary = "Get Executionplan and Steps",
        tags = { "query" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ExecutionplanStep.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/executionplan/{hash}/steps",
        produces = { "application/json" }
    )
    default ResponseEntity<List<ExecutionplanStep>> executionplanHashStepsGet(
        @Parameter(name = "hash", description = "Get steps of Executionplan", required = true, in = ParameterIn.PATH) @PathVariable("hash") String hash
    ) {
        return getDelegate().executionplanHashStepsGet(hash);
    }

}